Casting. Initialize NTRU public keys.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.17@3851 0a995953-517f-4c62-b348-104708d926b2

+++ Common/spot-on-external-address.cc
@@ -66,7 +66,7 @@ void spoton_external_address::slotFinished(void)
 	bytes.remove
 	  (0,
 	   bytes.indexOf("Current IP Address:") +
-	   qstrlen("Current IP Address:"));
+	   static_cast<int> (qstrlen("Current IP Address:")));
 
       indexOf = bytes.indexOf("<");
